Refactor starring projects for optimistic UI
Implement new project reducers that update the project stated
optimistically. If there's an error, the action is undone, and the lastest
state is restored.

Refactor projects listing, replacing the original reducer with the
'undo' support.

Implement a Saga method to handle server/connection errors and
restore the state by undoing the dispatched action.